+++ esp-lp-hal/src/lib.rs
@@ -55,6 +55,29 @@ cfg_if::cfg_if! {
     }
 }
 
+// ULP interrupt bitflags from:
+// https://github.com/espressif/esp-idf/blob/12f36a021f511cd4de41d3fffff146c5336ac1e7/components/ulp/ulp_riscv/ulp_core/ulp_riscv_interrupt.c#L16
+cfg_if::cfg_if! {
+    if #[cfg(any(esp32s2,esp32s3))] {
+        const ULP_RISCV_TIMER_INT                         : u32 = 1 << 0;   /* Internal Timer Interrupt */
+        const ULP_RISCV_EBREAK_ECALL_ILLEGAL_INSN_INT     : u32 = 1 << 1;   /* EBREAK, ECALL or Illegal instruction */
+        const ULP_RISCV_BUS_ERROR_INT                     : u32 = 1 << 2;   /* Bus Error (Unaligned Memory Access) */
+        const ULP_RISCV_PERIPHERAL_INTERRUPT              : u32 = 1 << 31;  /* RTC Peripheral Interrupt */
+        const ULP_RISCV_INTERNAL_INTERRUPT                : u32 = ULP_RISCV_TIMER_INT | ULP_RISCV_EBREAK_ECALL_ILLEGAL_INSN_INT | ULP_RISCV_BUS_ERROR_INT;
+    }
+}
+// ULP interrupt registers, for RTC_GPIO and SENS.
+cfg_if::cfg_if! {
+    if #[cfg(any(esp32s2,esp32s3))] {
+        const ULP_DR_REG_RTC_IO_BASE : u32 = 0xA400;
+        const ULP_DR_REG_SENS_BASE   : u32 = 0xC800;
+        const RTC_GPIO_STATUS_REG        : u32 = 0xA418;
+        const RTC_GPIO_STATUS_W1TC_REG   : u32 = 0xA420;
+        const SENS_SAR_COCPU_INT_ST_REG  : u32 = 0xC8F0;
+        const SENS_SAR_COCPU_INT_CLR_REG : u32 = 0xC8F4;
+    }
+}
+
 pub(crate) static mut CPU_CLOCK: u32 = LP_FAST_CLK_HZ;
 
 /// Wake up the HP core
@@ -112,41 +135,193 @@ loop:
 );
 
 #[cfg(any(esp32s2, esp32s3))]
+// Include the following macros which define custom RISCV R-Type instructions...
+//
+// # getq_insn rd, qs
+// Copies the value of Qx into a general purpose register rd
+//
+// # setq_insn qd, rs
+// Copies the value of general purpose register rs to Qx
+//
+// # retirq_insn
+// Copies the value of Q0 to CPU PC, and renables interrupts
+//
+// # maskirq_insn rd, rs
+// Copies the value of the register IRQ Mask to the register rd, and copies the value of register rs
+// to to IRQ mask.
+//
+// # waitirq_insn rd
+// Pause execution until any interrupt (masked or unmasked) becomes pending. Stores the pending IRQ
+// bitmask into register rd.
+global_asm!(include_str!("./ulp_riscv_interrupt_ops.S"));
+
+// Assembly containing the reset_vector and irq_vector instructions.
+#[cfg(any(esp32s2, esp32s3))]
+global_asm!(include_str!("./ulp_riscv_vectors.S"));
+
+#[cfg(any(esp32s2, esp32s3))]
 global_asm!(
     r#"
-	.section .text.vectors
-	.global irq_vector
-	.global reset_vector
+    .balign 0x10
+    .section .init
+    __start:
+        /* setup the stack pointer */
+        la sp, __stack_top
 
-/* The reset vector, jumps to startup code */
-reset_vector:
-	j __start
+        /* Custom instruction to un-mask the interrupts */
+        /* waitirq_insn zero */
+        maskirq_insn zero, zero
 
-/* Interrupt handler */
-.balign 16
-irq_vector:
-	ret
+        call ulp_riscv_rescue_from_monitor
+        call rust_main
 
-	.section .text
-
-__start:
-    /* setup the stack pointer */
-	la sp, __stack_top
-
-	call ulp_riscv_rescue_from_monitor
-	call rust_main
-	call ulp_riscv_halt
-loop:
-	j loop
-"#
+    loop:
+        j loop
+    "#
 );
 
-#[unsafe(link_section = ".init.rust")]
+#[cfg(esp32s2)]
+global_asm!(
+    r#"
+    /* Weakly-linked IRQ handler.
+    *  Simply clears the IRQ and returns.
+    */
+    ulp_interrupt_handler:
+        /* Register a0 == q1 == the bitmask of IRQs to be handled.
+        * 'bltz a0,1f' is checking if the MSB is set for an i32,
+        * indiciating that this is a ULP_RISCV_PERIPHERAL_INTERRUPT.
+        * All other interrupt types ('internal' interrupts) are ignored.
+        */
+        bltz    a0,1f
+        ret
+        1:
+        /* 1. Check for SENS_SAR peripheral interrupts.
+        * Load 0xD000 into a0,
+        * which will be used to
+        * read/write SENS_SAR registers.
+        */
+        lui     a0,0xd
+        /* Read interrupt flags from  
+        * SENS_SAR_COCPU_INT_ST_REG
+        * at address -1808(a0) = 0xC8F0.
+        */
+        lw      a1,-1808(a0)
+        /* If equal to zero then no
+        * interrupts have been raised,
+        * and we can skip clearing them.
+        */
+        beqz    a1,2f
+        /* Clear interrupts by writing
+        * to SENS_SAR_COCPU_INT_CLR_REG 
+        * at address -1804(a0) = 0xC8F4
+        */
+        sw      a1,-1804(a0)
+        2:
+        /* 2. Check for RTC_GPIO peripheral interrupts.
+        * Load 0xA000 into a0,
+        * which will be used to 
+        * read/write RTC_GPIO registers.
+        */
+        lui     a0,0xa
+        /* Read RTC_GPIO_STATUS_REG from +1048(a0) = 0xA418 */
+        lw      a1,1048(a0)
+        /* If no interrupt is raised, skip clearing them. */
+        beqz    a1,3f
+        /* Clear interrupts by writing
+        * to RTC_GPIO_STATUS_W1TC_REG
+        * at address +1056(a0) = 0xA420
+        */
+        sw      a1,1056(a0)
+        3:
+        ret 
+    "#
+);
+
+#[cfg(any(esp32s2, esp32s3))]
+#[unsafe(no_mangle)]
+unsafe extern "C" fn _ulp_riscv_interrupt_handler(q1: u32) {
+    // This interrupt handler is placeholder - it simply checks the interrupt flags, and clears
+    // them. This function is based on the ESP-IDF implementation found here:
+    // https://github.com/espressif/esp-idf/blob/12f36a021f511cd4de41d3fffff146c5336ac1e7/components/ulp/ulp_riscv/ulp_core/ulp_riscv_interrupt.c#L110
+
+    // ULP Internal Interrupts
+    if (q1 & ULP_RISCV_INTERNAL_INTERRUPT) > 0 {
+        // TODO
+    }
+
+    // External/Peripheral interrupts
+    if (q1 & ULP_RISCV_PERIPHERAL_INTERRUPT) > 0 {
+        // RTC Peripheral interrupts
+        let cocpu_int_st: u32 = unsafe { &*pac::SENS::PTR }.sar_cocpu_int_st().read().bits();
+
+        if cocpu_int_st > 0 {
+            // Clear the interrupt
+            unsafe { &*pac::SENS::PTR }
+                .sar_cocpu_int_clr()
+                .write(|w| unsafe { w.bits(cocpu_int_st) });
+        }
+
+        // RTC IO interrupts
+        let rtcio_int_st: u32 = unsafe { &*pac::RTC_IO::PTR }.status().read().bits();
+
+        if rtcio_int_st > 0 {
+            // Clear the interrupt
+            unsafe { &*pac::RTC_IO::PTR }
+                .status_w1tc()
+                .write(|w| unsafe { w.bits(rtcio_int_st) });
+        }
+    }
+}
+
+/// Enter a critical section (disable interrupts)
+#[cfg(any(esp32s2, esp32s3))]
+pub fn ulp_disable_interrupts() {
+    // Enter a critical section by disabling all interrupts
+    // This inline assembly construct uses the t0 register and is equivalent to:
+    // > li t0, 0x80000007
+    // > maskirq_insn(zero, t0) // Mask all interrupt bits
+    //
+    // The mask 0x80000007 represents:
+    //   Bit 31 - RTC peripheral interrupt
+    //   Bit 2  - Bus error
+    //   Bit 1  - Ebreak / Ecall / Illegal Instruction
+    //   Bit 0  - Internal Timer
+    //
+    unsafe {
+        core::arch::asm!("li t0, 0x80000007", ".word 0x0602e00b");
+    }
+}
+
+/// Exit a critical section (re-enable interrupts)
+#[cfg(any(esp32s2, esp32s3))]
+pub fn ulp_enable_interrupts() {
+    // Exit a critical section by enabling all interrupts
+    // This inline assembly construct is equivalent to:
+    // > maskirq_insn(zero, zero)
+    unsafe {
+        core::arch::asm!(".word 0x0600600b");
+    }
+}
+
+/// Wait for any (even unmasked) interrupt
+#[cfg(any(esp32s2, esp32s3))]
+pub fn ulp_waitirq() -> u32 {
+    // Wait for pending interrupt, return pending interrupt mask
+    // waitirq a0
+    let result: u32;
+    unsafe {
+        core::arch::asm!(".word 0x0800400B", out("a0") result);
+    }
+    result
+}
+
+/// Entry point to the ULP program
 #[unsafe(export_name = "rust_main")]
+#[unsafe(link_section = ".init")]
 unsafe extern "C" fn lp_core_startup() -> ! {
     unsafe {
         unsafe extern "Rust" {
-            fn main() -> !;
+            fn main();
         }
 
         #[cfg(esp32c6)]
@@ -160,12 +335,13 @@ unsafe extern "C" fn lp_core_startup() -> ! {
         }
 
         main();
+        ulp_riscv_halt();
     }
 }
 
 #[cfg(any(esp32s2, esp32s3))]
-#[unsafe(link_section = ".init.rust")]
 #[unsafe(no_mangle)]
+#[unsafe(link_section = ".init")]
 unsafe extern "C" fn ulp_riscv_rescue_from_monitor() {
     // Rescue RISC-V core from monitor state.
     unsafe { &*pac::RTC_CNTL::PTR }
@@ -173,15 +349,36 @@ unsafe extern "C" fn ulp_riscv_rescue_from_monitor() {
         .modify(|_, w| w.cocpu_done().clear_bit().cocpu_shut_reset_en().clear_bit());
 }
 
-#[cfg(any(esp32s2, esp32s3))]
+/// Stops the ULP core, called from itself.
 #[unsafe(link_section = ".init.rust")]
-#[unsafe(no_mangle)]
-unsafe extern "C" fn ulp_riscv_halt() {
-    unsafe { &*pac::RTC_CNTL::PTR }
-        .cocpu_ctrl()
-        .modify(|_, w| unsafe { w.cocpu_shut_2_clk_dis().bits(0x3f).cocpu_done().set_bit() });
+fn ulp_riscv_halt() -> ! {
+    #[cfg(any(esp32s2, esp32s3))]
+    {
+        unsafe { &*pac::RTC_CNTL::PTR }
+            .cocpu_ctrl()
+            .modify(|r, w| unsafe {
+                // Read the existing register value
+                let mut cocpu_ctl = r.bits();
+                // Set the value of COCPU_SHUT_2_CLK_DIS to 0x3F,
+                // as per the ESP-IDF source: https://github.com/espressif/esp-idf/blob/12f36a021f511cd4de41d3fffff146c5336ac1e7/components/ulp/ulp_riscv/ulp_core/ulp_riscv_utils.c#L29
+                cocpu_ctl &= 0xFFC03FFF;
+                cocpu_ctl |= 0x3F << 14;
+                // Write to the register.
+                w.bits(cocpu_ctl);
+                // Re-read the register
+                cocpu_ctl = r.bits();
+                // Set RTC_CNTL_COCPU_DONE and RTC_CNTL_COCPU_DONE
+                cocpu_ctl |= (1 << 22) | (1 << 25);
+                // Write to the register.
+                w.bits(cocpu_ctl);
+                w
+            });
+    }
 
+    // All chips will enter a no-op loop, when halting.
     loop {
-        riscv::asm::wfi();
+        unsafe {
+            core::arch::asm!("addi x0, x0, 0"); // no-op
+        }
     }
 }