+++ esp-hal/CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added experimental low-level clock control functionality via `esp_hal::clock::ll` (#5092)
 - Work queue support for ECC operations (#5084)
 - A mechanism to hook into linker scripts (#5116)
+- ESP32-S2,ESP32-S3: Add support for configuring the ULP Timer with `UlpCore.with_sleep_cycles(cycles : UlpCoreSleepCycles)` (#5134)
 - ESP32-S2, ESP32-S3: Add `UlpWakeupSource` so these chips can be woken up by ULP-generated interrupts. (#5132)
 
 ### Changed